Rename pseudo_types to type
